requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
jinja2==3.1.6
httpx==0.28.1
python-dotenv==1.1.0
python-whois==0.9.5
python-multipart==0.0.20
shodan==1.31.0
